+++ clock_domain_crossing/vhdl/cdc_sync_slv.vhd
@@ -28,7 +28,8 @@
 --                                       Clock2 ───┴───────────────┘
 -- 
 -- Revision:
---   0.01 - File Created
+--   [2024.02.20] - File Created
+--   [2024.03.09] - File renamed from cdc_sync to cdc_sync_slv
 --
 ----------------------------------------------------------------------------------
 
@@ -36,7 +37,7 @@ library IEEE;
 USE IEEE.STD_LOGIC_1164.ALL;
 use IEEE.NUMERIC_STD.ALL;
 
-entity cdc_sync is
+entity cdc_sync_slv is
 generic (
   REGS_STAGE : INTEGER :=  2;  -- Pipeline Registers
   DATA_WIDTH : INTEGER := 16
@@ -48,9 +49,9 @@ port(
   data_in   : IN  STD_LOGIC_VECTOR(DATA_WIDTH-1 DOWNTO 0);
   data_out  : OUT STD_LOGIC_VECTOR(DATA_WIDTH-1 DOWNTO 0)
 );
-end cdc_sync;
+end cdc_sync_slv;
 
-architecture rtl of cdc_sync is
+architecture rtl of cdc_sync_slv is
   
   TYPE T_arr IS ARRAY (natural range <>) OF STD_LOGIC_VECTOR(data_in'RANGE);
   SIGNAL regs : T_ARR(0 TO REGS_STAGE-1);