+++ memory/fifo/vhdl/axi_fifo_2clk.vhd
@@ -35,17 +35,23 @@ library IEEE;
 
 entity axi_fifo_2clk is
   generic (
-    ALMOST_FULL_OFFSET : NATURAL := 16;
-    FIFO_ADDR_WIDTH    : NATURAL := 9;
-    FIFO_DATA_WIDTH    : NATURAL := 16;
-    BRAM_TYPE          : STRING  := "block"
+    -- FIFO Parameters
+    MIN_SAMPLES_TO_READ : NATURAL :=  4;
+    ALMOST_FULL_OFFSET  : NATURAL := 16;
+    FIFO_ADDR_WIDTH     : NATURAL :=  9;
+    FIFO_DATA_WIDTH     : NATURAL := 16;
+    BRAM_TYPE           : STRING  := "block"
   );
   port (
+    
+    -- Write side
     i_clk    : in  STD_LOGIC;
     i_rst    : in  STD_LOGIC;
     i_tvalid : in  STD_LOGIC;
     i_tdata  : in  STD_LOGIC_VECTOR(FIFO_DATA_WIDTH - 1 downto 0);
     i_tready : out STD_LOGIC;
+    
+    -- Read side
     o_clk    : in  STD_LOGIC;
     o_rst    : in  STD_LOGIC;
     o_tvalid : out STD_LOGIC;
@@ -61,19 +67,21 @@ architecture rtl of axi_fifo_2clk is
 
   -- read_int will assert when either a read occurs or the output register is empty (and there is data in the shift register fifo)
   signal read_int : std_logic;
-  -- read will only assert when an actual 1read request occurs at the interface
-  --signal read : std_logic;
+
   -- Read side states
-  type Tstate is (ST_EMPTY, PRE_READ, READING);
-  signal read_state : Tstate := ST_EMPTY;
+  type Tstate is (IDLE, PRE_READ, READING);
+  signal read_state : Tstate := IDLE;
+  signal read_cnt   : unsigned(FIFO_ADDR_WIDTH - 1 downto 0);
 
   -- Addresses signals
-  signal wr_addr : unsigned(FIFO_ADDR_WIDTH - 1 downto 0);
-  signal rd_addr : unsigned(FIFO_ADDR_WIDTH - 1 downto 0);
+  signal wr_addr   : unsigned(FIFO_ADDR_WIDTH - 1 downto 0);
+  signal rd_addr   : unsigned(FIFO_ADDR_WIDTH - 1 downto 0);
+  signal diff_addr     : unsigned(FIFO_ADDR_WIDTH - 1 downto 0);
+  signal reg_diff_addr : unsigned(FIFO_ADDR_WIDTH - 1 downto 0);
 
   -- Empty and Full regs
-  signal empty_reg : std_logic := '1';
-  signal full_reg  : std_logic := '0';
+  signal reg_empty : std_logic := '1';
+  signal reg_full  : std_logic := '0';
   --signal full      : std_logic;
   signal empty     : std_logic;
   signal write     : std_logic;
@@ -91,7 +99,7 @@ architecture rtl of axi_fifo_2clk is
 
   -- Output Register
   signal reg_o_tvalid : std_logic := '0';
-  signal reg_tdata    : std_logic_vector(FIFO_DATA_WIDTH - 1 downto 0);
+  signal reg_o_tdata    : std_logic_vector(FIFO_DATA_WIDTH - 1 downto 0);
 
   -- Cross Domain Clock signals
   constant C_SYNC_STAGES : INTEGER := 2;
@@ -128,8 +136,10 @@ begin
     end if;
   end process;
   -------- Write Side - END --------
+  
+  
   -------- Cross Domain Clock Signals - BEGIN --------
-
+  
   CDC_IN_TO_OUT: process (o_clk)
   begin
     if rising_edge(o_clk) then
@@ -138,10 +148,10 @@ begin
         --elsif o_tready='1' then
       else
         cdc_wr_addr(0) <= wr_addr;
-        cdc_wr_addr(1 to cdc_wr_addr'LENGTH - 1) <= cdc_wr_addr(0 to cdc_wr_addr'LENGTH - 2);
+        cdc_wr_addr(1 to C_SYNC_STAGES - 1) <= cdc_wr_addr(0 to C_SYNC_STAGES - 2);
 
         cdc_write(0) <= write;
-        cdc_write(1 to cdc_write'LENGTH - 1) <= cdc_write(0 to cdc_write'LENGTH - 2);
+        cdc_write(1 to C_SYNC_STAGES - 1) <= cdc_write(0 to C_SYNC_STAGES - 2);
 
       end if;
     end if;
@@ -154,63 +164,84 @@ begin
         cdc_full <= (others => '0');
         --elsif o_tready='1' then
       else
-        cdc_full(0) <= full_reg;
-        cdc_full(1 to cdc_wr_addr'LENGTH - 1) <= cdc_full(0 to cdc_wr_addr'LENGTH - 2);
+        cdc_full(0) <= reg_full;
+        cdc_full(1 to C_SYNC_STAGES - 1) <= cdc_full(0 to C_SYNC_STAGES - 2);
       end if;
     end if;
   end process;
   -------- Cross Domain Clock Signals - END --------
+  
+  
   -------- Read Side - BEGIN --------
   read_int <= (not empty) and int_tready;
-  --read     <= o_tready and reg_o_tvalid;
 
   P_READ_DATA: process (o_clk)
   begin
     if rising_edge(o_clk) then
-      if read_state = PRE_READ or read_int = '1' then
-        int_tdata <= ram(TO_INTEGER(unsigned(rd_addr)));
+      if read_state = READING or read_int = '1' then
+        --int_tdata <= ram(TO_INTEGER(unsigned(rd_addr)));
+        reg_o_tdata <= ram(TO_INTEGER(unsigned(rd_addr)));
       end if;
     end if;
   end process;
+
+  diff_addr <= cdc_wr_addr(C_SYNC_STAGES-1)-rd_addr;
 
   P_READ_FSM: process (o_clk)
   begin
     if rising_edge(o_clk) then
       if o_rst = '1' then
-        read_state <= ST_EMPTY;
-        rd_addr <= (others => '0');
-        empty_reg <= '1';
+        read_state    <= IDLE;
+        reg_diff_addr <= (others => '0');
+        rd_addr       <= (others => '0');
+        read_cnt      <= (others => '0');
+        reg_empty     <= '1';
+        reg_o_tvalid  <= '0';
+
       else
 
         case read_state is
-          when ST_EMPTY =>
-            if cdc_write(cdc_write'LENGTH - 1) = '1' then
-              read_state <= PRE_READ;
-            end if;
 
-          when PRE_READ =>
-            read_state <= READING;
-            empty_reg <= '0';
-            rd_addr <= rd_addr + 1;
+          when IDLE =>
+            if diff_addr>MIN_SAMPLES_TO_READ-1 then
+              read_state    <= READING;
+              reg_diff_addr <= diff_addr;
+              reg_empty     <= '0';
+              -- reg_o_tdata   <= int_tdata;
+            end if;
 
           when READING =>
             if read_int = '1' then
-              if rd_addr = wr_addr then
-                empty_reg <= '1';
-                if cdc_write(cdc_write'LENGTH - 1) = '1' then
-                  read_state <= PRE_READ;
+              -- reg_o_tdata   <= int_tdata;
+
+              if read_cnt=reg_diff_addr-1 then
+                if diff_addr>MIN_SAMPLES_TO_READ-1 then
+                  reg_diff_addr <= diff_addr;
+                  reg_empty     <= '0';
+                  reg_o_tvalid  <= '1';
+                  read_cnt      <= (others => '0');
+                  rd_addr       <= rd_addr + 1;
                 else
-                  read_state <= ST_EMPTY;
+                  read_state   <= IDLE;
+                  reg_empty    <= '1';
+                  reg_o_tvalid <= '0';
+                  read_cnt     <= (others => '0');
+                  rd_addr      <= rd_addr;
                 end if;
               else
-                rd_addr <= rd_addr + 1;
+                reg_empty    <= '0';
+                reg_o_tvalid <= '1';
+                read_cnt     <= read_cnt + 1;
+                rd_addr      <= rd_addr  + 1;
               end if;
-            end if;
 
+            end if;
+  
           when others =>
-            read_state <= read_state;
-            rd_addr <= rd_addr;
-            empty_reg <= empty_reg;
+            read_state  <= read_state;
+            rd_addr     <= rd_addr;
+            read_cnt    <= read_cnt;
+            reg_empty   <= reg_empty;
 
         end case;
       end if;
@@ -224,40 +255,26 @@ begin
   begin
     if rising_edge(o_clk) then
       if o_rst = '1' then
-        full_reg <= '0';
-      elsif read_int = '1' and write = '0' then
-        full_reg <= '0';
-      elsif write = '1' and read_int = '0' and becoming_full = '1' then
-        full_reg <= '1';
+        reg_full <= '0';
+      elsif read_int = '1' then
+        reg_full <= '0';
+      elsif becoming_full = '1' then
+        reg_full <= '1';
       else
-        full_reg <= full_reg;
+        reg_full <= reg_full;
       end if;
     end if;
   end process;
 
-  empty <= empty_reg;
-  --full  <= full_reg;
-
-  -- Output registers
-  process (o_clk)
-  begin
-    if rising_edge(o_clk) then
-      if o_rst = '1' then
-        reg_o_tvalid <= '0';
-        reg_tdata <= (others => '0');
-      elsif int_tready = '1' then
-        reg_o_tvalid <= not empty;
-        reg_tdata <= int_tdata;
-      end if;
-    end if;
-  end process;
+  empty <= reg_empty;
 
   int_tready <= o_tready or (not reg_o_tvalid);
 
   -- Output ports
   o_tvalid <= reg_o_tvalid;
-  o_tdata  <= reg_tdata;
+  o_tdata  <= reg_o_tdata;
 
   -------- Read Side - END --------
+  
 end architecture;